requirements.txt
Babel==1.3
MySQL-python==1.2.5
PyYAML==3.11
Pygments==1.6
SQLAlchemy==0.9.7
argparse==1.2.1
bpython==0.13.1
cliff==1.7.0
cmd2==0.6.7
dnspython==1.12.0
ecdsa==0.11
httplib2==0.9
iso8601==0.1.10
jsonschema==2.4.0
netaddr==0.7.12
oslo.config==1.4.0
paramiko==1.15.1
pbr>=0.5.16,<0.6
pep8==1.5.7
prettytable==0.7.2
pycrypto==2.6.1
pyparsing==2.0.2
python-cinderclient==1.0.9
python-designateclient==1.1.0
python-keystoneclient==0.11.1
python-libraclient==2.0.0
python-neutronclient==2.3.4
python-novaclient==2.14.1
python-swiftclient==2.0.3
python-troveclient==1.0.5
pytz==2014.7
requests==2.4.1
simplejson==3.4.0
six==1.8.0
stevedore==1.0.0
wsgiref==0.1.2
